separated util file
