+++ actors/evm/src/interpreter/precompiles/bls12_381.rs
@@ -3,256 +3,32 @@ use super::PrecompileError;
 use super::PrecompileResult;
 use fil_actors_runtime::runtime::Runtime;
 use crate::interpreter::System;
+use super::bls_util::{
+    // Constants
+    G1_INPUT_LENGTH,
+    G1_ADD_INPUT_LENGTH,
+    G1_OUTPUT_LENGTH,
+    G1_MSM_INPUT_LENGTH,
+    G1_INPUT_ITEM_LENGTH,
+    G2_ADD_INPUT_LENGTH,
+    G2_INPUT_ITEM_LENGTH,
+    G2_OUTPUT_LENGTH,
+    G2_MSM_INPUT_LENGTH,
+    SCALAR_LENGTH,
+    NBITS,
 
-use blst::{
-    blst_p1, blst_p1_add_or_double_affine, blst_p1_affine, blst_p1_from_affine, blst_p1_to_affine, blst_p1_affine_in_g1, blst_fp, blst_p1_affine_on_curve, blst_fp_from_bendian, blst_bendian_from_fp, blst_scalar, blst_scalar_from_bendian, p1_affines, blst_p2, blst_p2_affine, blst_p2_add_or_double_affine, blst_p2_from_affine, blst_p2_to_affine, blst_p2_affine_on_curve, blst_p2_affine_in_g2, blst_fp2, p2_affines
+    // Functions
+    encode_g1_point,
+    encode_g2_point,
+    extract_g1_input,
+    extract_g2_input,
+    extract_scalar_input,
+
 };
 
-pub const G1_INPUT_LENGTH: usize = 128;
-pub const G1_ADD_INPUT_LENGTH: usize = G1_INPUT_LENGTH * 2;
-pub const G1_OUTPUT_LENGTH: usize = 128;
-pub const PADDED_FP_LENGTH: usize = 64;
-pub const PADDING_LENGTH: usize = 16;
-pub const G1_MSM_INPUT_LENGTH: usize = 160;
-pub const G1_INPUT_ITEM_LENGTH: usize = 128;
-pub const SCALAR_LENGTH: usize = 32;
-pub const NBITS: usize = 255; 
-pub const G2_ADD_INPUT_LENGTH: usize = 512;
-pub const G2_INPUT_ITEM_LENGTH: usize = 256;
-pub const G2_OUTPUT_LENGTH: usize = 256;
-pub const G2_MSM_INPUT_LENGTH: usize = 288;
-
-
-/// Encodes a G2 point in affine format into byte slice with padded elements.
-/// G2 points have two coordinates (x,y) where each coordinate is a complex number (real,imaginary)
-/// So we need to encode 4 field elements total: x.re, x.im, y.re, y.im
-pub(super) fn encode_g2_point(input: &blst_p2_affine) -> Vec<u8> {
-    // Create output buffer with space for all coordinates (4 * 64 bytes)
-    let mut out = vec![0u8; G2_OUTPUT_LENGTH];
-
-    // Encode x coordinate
-    // Real part (x.fp[0])
-    fp_to_bytes(&mut out[..PADDED_FP_LENGTH], &input.x.fp[0]);
-    // Imaginary part (x.fp[1]) 
-    fp_to_bytes(
-        &mut out[PADDED_FP_LENGTH..2 * PADDED_FP_LENGTH],
-        &input.x.fp[1],
-    );
-
-    // Encode y coordinate
-    // Real part (y.fp[0])
-    fp_to_bytes(
-        &mut out[2 * PADDED_FP_LENGTH..3 * PADDED_FP_LENGTH],
-        &input.y.fp[0],
-    );
-    // Imaginary part (y.fp[1])
-    fp_to_bytes(
-        &mut out[3 * PADDED_FP_LENGTH..4 * PADDED_FP_LENGTH],
-        &input.y.fp[1],
-    );
-
-    out
-}
-
-/// Convert field elements from byte slices into a `blst_p2_affine` point.
-/// Takes four 48-byte arrays representing:
-/// - x1: real part of x coordinate
-/// - x2: imaginary part of x coordinate
-/// - y1: real part of y coordinate
-/// - y2: imaginary part of y coordinate
-pub(super) fn decode_and_check_g2(
-    x1: &[u8; 48], // x.re
-    x2: &[u8; 48], // x.im
-    y1: &[u8; 48], // y.re
-    y2: &[u8; 48], // y.im
-) -> Result<blst_p2_affine, PrecompileError> {
-    Ok(blst_p2_affine {
-        // Create x coordinate as complex number
-        x: check_canonical_fp2(x1, x2)?,
-        // Create y coordinate as complex number
-        y: check_canonical_fp2(y1, y2)?,
-    })
-}
-
-/// Helper function to create and validate an Fp2 element from two Fp elements
-fn check_canonical_fp2(
-    input_1: &[u8; 48],
-    input_2: &[u8; 48],
-) -> Result<blst_fp2, PrecompileError> {
-    let fp_1 = fp_from_bendian(input_1)?;
-    let fp_2 = fp_from_bendian(input_2)?;
-
-    let fp2 = blst_fp2 { fp: [fp_1, fp_2] };
-
-    Ok(fp2)
-}
-
-
-/// Extracts a G2 point in Affine format from a 256 byte slice representation.
-///
-/// **Note**: This function will perform a G2 subgroup check if `subgroup_check` is set to `true`.
-/// 
-/// Subgroup checks are required for:
-/// - Scalar multiplication
-/// - Multi-scalar multiplication (MSM)
-/// - Pairing operations
-///
-/// But not required for:
-/// - Point addition
-/// - Point negation
-pub(super) fn extract_g2_input(
-    input: &[u8],
-    subgroup_check: bool,
-) -> Result<blst_p2_affine, PrecompileError> {
-    // Check input length (256 bytes = 4 * 64 bytes for x.re, x.im, y.re, y.im)
-    if input.len() != G2_INPUT_ITEM_LENGTH {
-        return Err(PrecompileError::IncorrectInputSize);
-    }
-
-    // Extract the four field elements (removing padding)
-    let x_re = remove_padding(&input[..PADDED_FP_LENGTH])?;
-    let x_im = remove_padding(&input[PADDED_FP_LENGTH..2 * PADDED_FP_LENGTH])?;
-    let y_re = remove_padding(&input[2 * PADDED_FP_LENGTH..3 * PADDED_FP_LENGTH])?;
-    let y_im = remove_padding(&input[3 * PADDED_FP_LENGTH..4 * PADDED_FP_LENGTH])?;
-
-    // Convert bytes to point
-    let point = decode_and_check_g2(x_re, x_im, y_re, y_im)?;
-
-    if subgroup_check {
-        // Subgroup check (more expensive but required for certain operations)
-        // Verifies that the point has the correct order and is in G2
-        // SAFETY: point is properly initialized above
-        unsafe {
-            if !blst_p2_affine_in_g2(&point) {
-                return Err(PrecompileError::InvalidInput);
-            }
-        }
-    } else {
-        // Basic curve check (less expensive, sufficient for addition)
-        // Only verifies that the point is on the curve
-        // SAFETY: point is properly initialized above
-        if unsafe { !blst_p2_affine_on_curve(&point) } {
-            return Err(PrecompileError::InvalidInput);
-        }
-    }
-
-    Ok(point)
-}
-
-/// https://eips.ethereum.org/EIPS/eip-2537
-/// Encodes a single finite field element into byte slice with padding.
-pub(super) fn fp_to_bytes(out: &mut [u8], input: *const blst_fp) {
-    if out.len() != PADDED_FP_LENGTH {
-        return;
-    }
-    let (padding, rest) = out.split_at_mut(PADDING_LENGTH);
-    padding.fill(0);
-    // SAFETY: Out length is checked previously, `input` is a blst value.
-    unsafe { blst_bendian_from_fp(rest.as_mut_ptr(), input) };
-}
-
-/// Checks whether or not the input represents a canonical field element, returning the field
-/// element if successful.
-fn fp_from_bendian(bytes: &[u8; 48]) -> Result<blst_fp, PrecompileError> {
-    let mut fp = blst_fp::default();
-    unsafe {
-        // This performs the check for canonical field elements
-        blst_fp_from_bendian(&mut fp, bytes.as_ptr());
-    }
-    Ok(fp)
-}
-
-
-/// Extracts a scalar value from a 32-byte input.
-/// 
-/// According to EIP-2537, the scalar input:
-/// - Must be exactly 32 bytes
-/// - Is interpreted as a big-endian integer
-/// - Is not required to be less than the curve order
-/// 
-/// Returns a Result containing either the scalar value or a PrecompileError
-pub(super) fn extract_scalar_input(input: &[u8]) -> Result<blst_scalar, PrecompileError> {
-    // Check input length
-    if input.len() != SCALAR_LENGTH {
-        return Err(PrecompileError::IncorrectInputSize);
-    }
-
-    let mut scalar = blst_scalar::default();
-    
-    // Convert from big-endian bytes to scalar
-    // SAFETY: Input length is checked above and scalar is properly initialized
-    unsafe {
-        blst_scalar_from_bendian(&mut scalar, input.as_ptr());
-    }
-
-    Ok(scalar)
-}
-
-/// Returns a `blst_p1_affine` from the provided byte slices, which represent the x and y
-/// affine coordinates of the point.
-///
-/// If the x or y coordinate do not represent a canonical field element, an error is returned.
-///
-/// See [fp_from_bendian] for more information.
-fn decode_and_check_g1(
-    x_bytes: &[u8; 48],
-    y_bytes: &[u8; 48],
-) -> Result<blst_p1_affine, PrecompileError> {
-    Ok(blst_p1_affine {
-        x: fp_from_bendian(x_bytes)?,
-        y: fp_from_bendian(y_bytes)?,
-    })
-}
-/// Extracts a G1 point in Affine format from a 128 byte slice representation.
-fn extract_g1_input(input: &[u8], subgroup_check: bool) -> Result<blst_p1_affine, PrecompileError> {
-    if input.len() != G1_INPUT_LENGTH {
-        return Err(PrecompileError::IncorrectInputSize);
-    }
-
-    // Split input and remove padding for x and y coordinates
-    let x_bytes = remove_padding(&input[..PADDED_FP_LENGTH])?;
-    let y_bytes = remove_padding(&input[PADDED_FP_LENGTH..G1_INPUT_LENGTH])?;
- 
-    let point = decode_and_check_g1(x_bytes, y_bytes)?;
-
-    // Check if point is on curve (no subgroup check needed for addition)
-    if subgroup_check {
-        if unsafe { !blst_p1_affine_in_g1(&point) } {
-            return Err(PrecompileError::InvalidInput);
-        }
-    }
-    else{
-        unsafe {
-            if !blst_p1_affine_on_curve(&point) {
-                return Err(PrecompileError::InvalidInput);
-            }
-        }
-    }
-    Ok(point)
-}
-
-/// Removes zeros with which the precompile inputs are left padded to 64 bytes.
-fn remove_padding(input: &[u8]) -> Result<&[u8; 48], PrecompileError> {
-    if input.len() != PADDED_FP_LENGTH {
-        return Err(PrecompileError::IncorrectInputSize);
-    }
-    let (padding, unpadded) = input.split_at(PADDING_LENGTH);
-    if !padding.iter().all(|&x| x == 0) {
-        return Err(PrecompileError::InvalidInput);
-    }
-    unpadded.try_into().map_err(|_| PrecompileError::IncorrectInputSize)
-}
-
-/// Encodes a G1 point in affine format into byte slice with padded elements.
-fn encode_g1_point(input: *const blst_p1_affine) -> Vec<u8> {
-    let mut out = vec![0u8; G1_OUTPUT_LENGTH];
-    // SAFETY: Out comes from fixed length array, input is a blst value.
-    unsafe {
-        fp_to_bytes(&mut out[..PADDED_FP_LENGTH], &(*input).x);
-        fp_to_bytes(&mut out[PADDED_FP_LENGTH..], &(*input).y);
-    }
-    out.into()
-}
+use blst::{
+    blst_p1, blst_p1_add_or_double_affine, blst_p1_affine, blst_p1_from_affine, blst_p1_to_affine, p1_affines, blst_p2, blst_p2_affine, blst_p2_add_or_double_affine, blst_p2_from_affine, blst_p2_to_affine, p2_affines
+};
 
 /// BLS12_G1ADD precompile
 /// Implements G1 point addition according to EIP-2537
@@ -551,24 +327,6 @@ mod tests {
         assert!(matches!(res, Err(PrecompileError::IncorrectInputSize)));
     }
 
-    #[test]
-    fn test_fp_conversion() {
-        // Test fp_to_bytes and fp_from_bendian
-        let test_bytes: [u8; 48] = [
-            0x11, 0x22, 0x33, 0x44, 0x55, 0x66, 0x77, 0x88,
-            0x99, 0xaa, 0xbb, 0xcc, 0xdd, 0xee, 0xff, 0x00,
-            0x11, 0x22, 0x33, 0x44, 0x55, 0x66, 0x77, 0x88,
-            0x99, 0xaa, 0xbb, 0xcc, 0xdd, 0xee, 0xff, 0x00,
-            0x11, 0x22, 0x33, 0x44, 0x55, 0x66, 0x77, 0x88,
-            0x99, 0xaa, 0xbb, 0xcc, 0xdd, 0xee, 0xff, 0x00,
-        ];
-
-        // Test roundtrip conversion
-        let fp = fp_from_bendian(&test_bytes).unwrap();
-        let mut output = [0u8; 48];
-        fp_to_bytes(&mut output, &fp);
-        assert_eq!(test_bytes, output);
-    }
     #[test]
     fn test_g1_msm_success() {
         let rt = MockRuntime::default();
@@ -871,5 +629,135 @@ mod tests {
             "G2 MSM test case 3 (g2 * 1) failed.\nGot: {}\nExpected: {}",
             hex::encode(&res), hex::encode(&expected3)
         );
+        // Test case 4: p2 * 1 (identity operation)
+        let input4 = hex!(
+            "00000000000000000000000000000000103121a2ceaae586d240843a398967325f8eb5a93e8fea99b62b9f88d8556c80dd726a4b30e84a36eeabaf3592937f27\
+            00000000000000000000000000000000086b990f3da2aeac0a36143b7d7c824428215140db1bb859338764cb58458f081d92664f9053b50b3fbd2e4723121b68\
+            000000000000000000000000000000000f9e7ba9a86a8f7624aa2b42dcc8772e1af4ae115685e60abc2c9b90242167acef3d0be4050bf935eed7c3b6fc7ba77e\
+            000000000000000000000000000000000d22c3652d0dc6f0fc9316e14268477c2049ef772e852108d269d9c38dba1d4802e8dae479818184c08f9a569d878451\
+            0000000000000000000000000000000000000000000000000000000000000001"
+        );
+
+        let expected4 = hex!(
+            "00000000000000000000000000000000103121a2ceaae586d240843a398967325f8eb5a93e8fea99b62b9f88d8556c80dd726a4b30e84a36eeabaf3592937f27\
+            00000000000000000000000000000000086b990f3da2aeac0a36143b7d7c824428215140db1bb859338764cb58458f081d92664f9053b50b3fbd2e4723121b68\
+            000000000000000000000000000000000f9e7ba9a86a8f7624aa2b42dcc8772e1af4ae115685e60abc2c9b90242167acef3d0be4050bf935eed7c3b6fc7ba77e\
+            000000000000000000000000000000000d22c3652d0dc6f0fc9316e14268477c2049ef772e852108d269d9c38dba1d4802e8dae479818184c08f9a569d878451"
+        );
+
+        let res = bls12_g2msm(&mut system, &input4, PrecompileContext::default()).unwrap();
+        assert_eq!(res, expected4,
+            "G2 MSM test case 4 (p2 * 1) failed.\nGot: {}\nExpected: {}",
+            hex::encode(&res), hex::encode(&expected4)
+        );
+
+        // Test case 5: g2 * 0 (multiplication by zero)
+        let input5 = hex!(
+            "00000000000000000000000000000000024aa2b2f08f0a91260805272dc51051c6e47ad4fa403b02b4510b647ae3d1770bac0326a805bbefd48056c8c121bdb8\
+            0000000000000000000000000000000013e02b6052719f607dacd3a088274f65596bd0d09920b61ab5da61bbdc7f5049334cf11213945d57e5ac7d055d042b7e\
+            000000000000000000000000000000000ce5d527727d6e118cc9cdc6da2e351aadfd9baa8cbdd3a76d429a695160d12c923ac9cc3baca289e193548608b82801\
+            000000000000000000000000000000000606c4a02ea734cc32acd2b02bc28b99cb3e287e85a763af267492ab572e99ab3f370d275cec1da1aaa9075ff05f79be\
+            0000000000000000000000000000000000000000000000000000000000000000"
+        );
+
+        let expected5 = hex!(
+            "00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000\
+            00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000\
+            00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000\
+            00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000"
+        );
+
+        let res = bls12_g2msm(&mut system, &input5, PrecompileContext::default()).unwrap();
+        assert_eq!(res, expected5,
+            "G2 MSM test case 5 (g2 * 0) failed.\nGot: {}\nExpected: {}",
+            hex::encode(&res), hex::encode(&expected5)
+        );
+
+        // // Test case 6: p2 * 0 (multiplication by zero)
+        // let input6 = hex!(
+        //     "00000000000000000000000000000000103121a2ceaae586d240843a398967325f8eb5a93e8fea99b62b9f88d8556c80dd726a4b30e84a36eeabaf3592937f27\
+        //     00000000000000000000000000000000086b990f3da2aeac0a36143b7d7c824428215140db1bb859338764cb58458f081d92664f9053b50b3fbd2e4723121b68\
+        //     000000000000000000000000000000000f9e7ba9a86a8f7624aa2b42dcc8772e1af4ae115685e60abc2c9b90242167acef3d0be4050bf935eed7c3b6fc7ba77e\
+        //     000000000000000000000000000000000d22c3652d0dc6f0fc9316e14268477c2049ef772e852108d269d9c38dba1d4802e8dae479818184c08f9a569d878451\
+        //     0000000000000000000000000000000000000000000000000000000000000000"
+        // );
+
+        // let expected6 = hex!(
+        //     "00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000\
+        //     00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000\
+        //     00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000\
+        //     00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000"
+        // );
+
+        // let res = bls12_g2msm(&mut system, &input6, PrecompileContext::default()).unwrap();
+        // assert_eq!(res, expected6,
+        //     "G2 MSM test case 6 (p2 * 0) failed.\nGot: {}\nExpected: {}",
+        //     hex::encode(&res), hex::encode(&expected6)
+        // );
+
+        // // Test case 7: infinity * x
+        // let input7 = hex!(
+        //     "00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000\
+        //     00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000\
+        //     00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000\
+        //     00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000\
+        //     0000000000000000000000000000000000000000000000000000000000000011"
+        // );
+
+        // let expected7 = hex!(
+        //     "00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000\
+        //     00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000\
+        //     00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000\
+        //     00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000"
+        // );
+
+        // let res = bls12_g2msm(&mut system, &input7, PrecompileContext::default()).unwrap();
+        // assert_eq!(res, expected7,
+        //     "G2 MSM test case 7 (infinity * x) failed.\nGot: {}\nExpected: {}",
+        //     hex::encode(&res), hex::encode(&expected7)
+        // );
+        // // Test case 8: Random scalar * g2
+        // let input8 = hex!(
+        //     "00000000000000000000000000000000024aa2b2f08f0a91260805272dc51051c6e47ad4fa403b02b4510b647ae3d1770bac0326a805bbefd48056c8c121bdb8\
+        //     0000000000000000000000000000000013e02b6052719f607dacd3a088274f65596bd0d09920b61ab5da61bbdc7f5049334cf11213945d57e5ac7d055d042b7e\
+        //     000000000000000000000000000000000ce5d527727d6e118cc9cdc6da2e351aadfd9baa8cbdd3a76d429a695160d12c923ac9cc3baca289e193548608b82801\
+        //     000000000000000000000000000000000606c4a02ea734cc32acd2b02bc28b99cb3e287e85a763af267492ab572e99ab3f370d275cec1da1aaa9075ff05f79be\
+        //     263dbd792f5b1be47ed85f8938c0f29586af0d3ac7b977f21c278fe1462040e3"
+        // );
+
+        // let expected8 = hex!(
+        //     "0000000000000000000000000000000014856c22d8cdb2967c720e963eedc999e738373b14172f06fc915769d3cc5ab7ae0a1b9c38f48b5585fb09d4bd2733bb\
+        //     000000000000000000000000000000000c400b70f6f8cd35648f5c126cce5417f3be4d8eefbd42ceb4286a14df7e03135313fe5845e3a575faab3e8b949d2488\
+        //     00000000000000000000000000000000149a0aacc34beba2beb2f2a19a440166e76e373194714f108e4ab1c3fd331e80f4e73e6b9ea65fe3ec96d7136de81544\
+        //     000000000000000000000000000000000e4622fef26bdb9b1e8ef6591a7cc99f5b73164500c1ee224b6a761e676b8799b09a3fd4fa7e242645cc1a34708285e4"
+        // );
+
+        // let res = bls12_g2msm(&mut system, &input8, PrecompileContext::default()).unwrap();
+        // assert_eq!(res, expected8,
+        //     "G2 MSM test case 8 (random scalar * g2) failed.\nGot: {}\nExpected: {}",
+        //     hex::encode(&res), hex::encode(&expected8)
+        // );
+
+        // // Test case 9: Random scalar * p2
+        // let input9 = hex!(
+        //     "00000000000000000000000000000000103121a2ceaae586d240843a398967325f8eb5a93e8fea99b62b9f88d8556c80dd726a4b30e84a36eeabaf3592937f27\
+        //     00000000000000000000000000000000086b990f3da2aeac0a36143b7d7c824428215140db1bb859338764cb58458f081d92664f9053b50b3fbd2e4723121b68\
+        //     000000000000000000000000000000000f9e7ba9a86a8f7624aa2b42dcc8772e1af4ae115685e60abc2c9b90242167acef3d0be4050bf935eed7c3b6fc7ba77e\
+        //     000000000000000000000000000000000d22c3652d0dc6f0fc9316e14268477c2049ef772e852108d269d9c38dba1d4802e8dae479818184c08f9a569d878451\
+        //     263dbd792f5b1be47ed85f8938c0f29586af0d3ac7b977f21c278fe1462040e3"
+        // );
+
+        // let expected9 = hex!(
+        //     "00000000000000000000000000000000036074dcbbd0e987531bfe0e45ddfbe09fd015665990ee0c352e8e403fe6af971d8f42141970d9ab14b4dd04874409e6\
+        //     00000000000000000000000000000000019705637f24ba2f398f32c3a3e20d6a1cd0fd63e6f8f071cf603a8334f255744927e7bfdfdb18519e019c49ff6e9145\
+        //     00000000000000000000000000000000008e74fcff4c4278c9accfb60809ed69bbcbe3d6213ef2304e078d15ec7d6decb4f462b24b8e7cc38cc11b6f2c9e0486\
+        //     0000000000000000000000000000000001331d40100f38c1070afd832445881b47cf4d63894666d9907c85ac66604aab5ad329980938cc3c167ccc5b6bc1b8f30"
+        // );
+
+        // let res = bls12_g2msm(&mut system, &input9, PrecompileContext::default()).unwrap();
+        // assert_eq!(res, expected9,
+        //     "G2 MSM test case 9 (random scalar * p2) failed.\nGot: {}\nExpected: {}",
+        //     hex::encode(&res), hex::encode(&expected9)
+        // );
     }
 }
